avatar_3d_pipeline/requirements.txt
# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Numerical and vision
numpy>=1.26.0
scipy>=1.12.0
pillow>=10.3.0
opencv-python>=4.9.0
trimesh>=4.4.0
open3d>=0.18.0

# Deep learning stack (install torch with CUDA 12.1 index URL per README/instructions)
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
torch-geometric>=2.5.0

# 3D and model ecosystem
pytorch3d>=0.7.7
huggingface_hub>=0.24.0
gdown>=5.2.0
requests>=2.32.0

# Utilities
tqdm>=4.66.0
